Lab 5 prompt
Write a program that reads integers until EOF is entered. After each integer is entered, your program will display:

The binary representation of the integer (8 binary digits)
The hexadecimal representation of the integer (2 hex digits)
The character representation of the integer, separated by a tab character
If the user enters a negative value, your program should alert the user to try again, and ask for an integer again. If the value entered is greater than the range of ASCII (255), then your program should assume the value entered is 255.
Note: the final input of this example is EOF (CTRL + D). 
Another note: special ASCII values are not readable (looks like a box) when printed as a character, such as EOF (4). This is expected.

Follow the output format of the example. In other words, all strings should match (e.g. "Enter a number").